[QuazalServer/MitmDNS] - Improves the Spark Service to allow to host and play parties (very WIP), fixes memory leaks (again), fixes bugs, adds up missing licence added on the upstream and adds a new DNS entry.
Massive thanks to Halvors on discord and IDA researches that have helped me out with Spark.

Still some bits to figure out, and the service lacks any meaningful parties info/player storage which needs to be there for any decent hosting.

+++ Servers/QuazalServer/RDVServices/RMC/RMCPacket.cs
@@ -110,7 +110,7 @@ namespace QuazalServer.RDVServices.RMC
         {
             MemoryStream packetData = new();
 
-            if ((ushort)proto < 0x7F)
+            if (proto < 0x7F)
             {
                 // request has 0x80 flag
                 uint protoIdent = (uint)proto | (isRequest ? 0x80u : 0x0u);
@@ -120,7 +120,7 @@ namespace QuazalServer.RDVServices.RMC
             {
                 uint protoIdent = 0x7Fu | (isRequest ? 0x80u : 0x0u);
                 Helper.WriteU8(packetData, (byte)protoIdent);
-                Helper.WriteU16(packetData, (ushort)proto);
+                Helper.WriteU16(packetData, proto);
             }
 
             byte[] buff = Array.Empty<byte>();
